fix: sst global import
