+++ src/components/index.ts
@@ -1,1 +1,2 @@
 export * as ix from "./ix/index.js";
+export * from "./setup-components.js";
