+++ src/index.ts
@@ -1,3 +1,3 @@
 export * from "./deployConfig.js";
-export * from "./lib/sst/component-defaults.js";
+export * from "./components/setup-components.js";
 export * from "./lib/proxy/fetch.js";
